// src/types.ts
export enum Role {
  Student = 'Student',
  HallHead = 'Hall Head',
  SubjectTeacher = 'Subject Teacher',
  Finance = 'Finance Staff',
  Security = 'Security',
  Admin = 'System Admin',
}

export enum ClearanceStatus {
  Pending = 'Pending',
  Approved = 'Approved',
  Flagged = 'Flagged',
}

export interface User {
  id: string;
  email: string;
  name: string;
  role: Role;
  password?: string;
}

export interface Book {
  id: string;
  title: string;
  returned: boolean;
  category?: 'Book' | 'Other Item';
  imageUrl?: string;
  issuedDate?: string;
  issuerComment?: string;
  issuerId?: string;
}

export interface Fine {
  id: string;
  reason: string;
  amount: number;
  paid: boolean;
  receiptUrl?: string;
}

export interface SubjectClearance {
  id: string;
  subjectName: string;
  teacherName: string;
  teacherId: string;
  status: ClearanceStatus;
  comment: string;
  updatedAt: string;
}

export interface Approval {
  id: string;
  department: 'HallHead' | 'Finance' | 'Security';
  status: ClearanceStatus;
  approverName: string;
  comment: string;
  updatedAt: string;
}

export interface Student extends User {
  role: Role.Student;
  studentId: string; // e.g., 'ALA2025-001'
  hall: string;
  room: string;
  hallHeadName: string;
  books: Book[];
  fines: Fine[];
  subjectClearances: SubjectClearance[];
  approvals: Approval[];
}

export interface Staff extends User {
  role: Role.HallHead | Role.SubjectTeacher | Role.Finance | Role.Security | Role.Admin;
}

export interface Log {
  id: string;
  loggerId: string; // staff email
  role: Role;
  action: string;
  timestamp: string;
}